Add tests to solved problems
