+++ 2021/day04-giant-squid/problem1.py
@@ -1,6 +1,8 @@
-with open("input.txt") as f:
+with open("2021/day04-giant-squid/input.txt") as f:
     input = [line.replace("\n", "") for line in f]
     numbers_drawn = [int(i) for i in input[0].split(",")]
+
+def makeBoards(input):
     input = input[2:]
 
     boards = []
@@ -12,6 +14,8 @@ with open("input.txt") as f:
         else: 
             curr_board.append([int(i) for i in line.split()])
     boards.append(curr_board) #edge case of last board
+
+    return boards
 
 def checkBoard(board):
     for row in range(5):
@@ -26,14 +30,13 @@ def checkBoard(board):
             return True
     return False
 
-def markBoard(num_called):
-    global boards
+def markBoard(num_called, boards):
     for board in boards:
         for y in range(5):
             for x in range(5):
                 if board[y][x] == num_called:
                     board[y][x] = -1
-
+    return boards
 
 def getUnmarked(board):
     unmarked = []
@@ -45,12 +48,11 @@ def getUnmarked(board):
     return unmarked
 
 
-def playBingo(numbers):
-    global boards
+def playBingo(numbers, boards):
     final_score = 0
 
     for number in numbers:
-        markBoard(number)
+        boards = markBoard(number, boards)
         for board in boards:
             if checkBoard(board):
                 numbers_left =getUnmarked(board)                    
@@ -58,5 +60,6 @@ def playBingo(numbers):
                 final_score *= number
                 return final_score
 
-print(playBingo(numbers_drawn))
+boards = makeBoards(input)
+print(playBingo(numbers_drawn, boards))
