+++ 2021/day04-giant-squid/problem2.py
@@ -1,6 +1,8 @@
-with open("input.txt") as f:
+with open("2021/day04-giant-squid/input.txt") as f:
     input = [line.replace("\n", "") for line in f]
     numbers_drawn = [int(i) for i in input[0].split(",")]
+
+def makeBoards(input):
     input = input[2:]
 
     boards = []
@@ -13,6 +15,8 @@ with open("input.txt") as f:
             curr_board.append([int(i) for i in line.split()])
     boards.append(curr_board) #edge case of last board
 
+    return boards
+
 def checkBoard(board):
     for row in range(5):
         if sum(board[row]) == -5:
@@ -23,14 +27,13 @@ def checkBoard(board):
             return True
     return False
 
-def markBoard(num_called):
-    global boards
+def markBoard(num_called, boards):
     for board in boards:
         for y in range(5):
             for x in range(5):
                 if board[y][x] == num_called:
                     board[y][x] = -1
-
+    return boards
 
 def getUnmarked(board):
     unmarked = []
@@ -41,14 +44,13 @@ def getUnmarked(board):
 
     return unmarked
 
-def playBingo(numbers):
-    global boards
+def playBingo(numbers, boards):
     final_score = 0
 
     for number in numbers:
         board_won = False
         winning_boards = []
-        markBoard(number)
+        boards = markBoard(number, boards)
         
         for i in range(len(boards)):
             if checkBoard(boards[i]) and len(boards) > 1:
@@ -66,4 +68,5 @@ def playBingo(numbers):
                 if i not in winning_boards:
                     boards.append(boards_copy[i])
 
-print(playBingo(numbers_drawn))
+boards = makeBoards(input)
+print(playBingo(numbers_drawn, boards))
