+++ 2021/day07-the-treachery-of-whales/day07_problem1.py
@@ -1,0 +1,18 @@
+with open("2021/day07-the-treachery-of-whales/input.txt") as f:
+    crab_positions = [int(x) for x in f.readline().rstrip('\n').split(",")]
+
+def calc_least_fuel(crab_positions):
+    mode_pos = max(set(crab_positions), key=crab_positions.count)
+    mean_pos = int((sum(crab_positions) / len(crab_positions)))
+
+    fuel_spent = sum(crab_positions)
+
+    for i in range(min(mean_pos, mode_pos), max(mean_pos, mode_pos) + 1): 
+        curr_fuel_spent = 0
+        for pos in crab_positions:
+            curr_fuel_spent += abs((pos - i))
+        if curr_fuel_spent < fuel_spent:
+            fuel_spent = curr_fuel_spent
+    return fuel_spent
+
+print(calc_least_fuel(crab_positions))
