--- 2021/day07-the-treachery-of-whales/problem1.py
@@ -1,16 +1,0 @@
-with open("input.txt") as f:
-    crab_positions = [int(x) for x in f.readline().rstrip('\n').split(",")]
-
-mode_pos = max(set(crab_positions), key=crab_positions.count)
-mean_pos = int((sum(crab_positions) / len(crab_positions)))
-
-fuel_spent = sum(crab_positions)
-
-for i in range(min(mean_pos, mode_pos), max(mean_pos, mode_pos) + 1): 
-    curr_fuel_spent = 0
-    for pos in crab_positions:
-        curr_fuel_spent += abs((pos - i))
-    if curr_fuel_spent < fuel_spent:
-        fuel_spent = curr_fuel_spent
-
-print(fuel_spent)
