+++ 2021/day08-seven-segment-search/test_input.txt
