+++ 2021/day09-smoke-basin/test_input.txt
