+++ 2021/day16-packet-decoder/day16_problem1.py
@@ -1,31 +1,37 @@
 from collections import namedtuple
 
 packet = namedtuple("packet", ("version", "type_id", "value", "subpackets"))
-with open("input.txt") as f:
-    input = f.readline().rstrip("\n")
-    
-hex_to_bin = {
-    "0": "0000",
-    "1": "0001",
-    "2": "0010",
-    "3": "0011",
-    "4": "0100",
-    "5": "0101",
-    "6": "0110",
-    "7": "0111",
-    "8": "1000",
-    "9": "1001",
-    "A": "1010",
-    "B": "1011",
-    "C": "1100",
-    "D": "1101",
-    "E": "1110",
-    "F": "1111"
-}
 
-transmission = ""
-for hex in input:
-    transmission += hex_to_bin[hex]
+def convert_to_hex(file):
+    if file == "input":
+        with open(f"2021/day16-packet-decoder/{file}.txt") as f:
+            input = f.readline().rstrip("\n")
+    else:
+        input = file
+        
+    hex_to_bin = {
+        "0": "0000",
+        "1": "0001",
+        "2": "0010",
+        "3": "0011",
+        "4": "0100",
+        "5": "0101",
+        "6": "0110",
+        "7": "0111",
+        "8": "1000",
+        "9": "1001",
+        "A": "1010",
+        "B": "1011",
+        "C": "1100",
+        "D": "1101",
+        "E": "1110",
+        "F": "1111"
+    }
+
+    transmission = ""
+    for hex in input:
+        transmission += hex_to_bin[hex]
+    return transmission
 
 def decodeTransmission(transmission, i):
     version = int(transmission[i:i + 3], 2)
@@ -85,5 +91,9 @@ def sumVersions(packet):
         total += sumVersions(subpacket)
     return total
 
-packet = decodeTransmission(transmission, 0)[0]
-print(sumVersions(packet))
+def get_answer(file):
+    transmission = convert_to_hex(file)
+    packet = decodeTransmission(transmission, 0)[0]
+    return sumVersions(packet)
+    
+print(get_answer("input"))
