--- 2021/day16-packet-decoder/test_input.txt
@@ -1,1 +1,0 @@
-D2FE28
