--- 2021/day16-packet-decoder/test_input2.txt
@@ -1,1 +1,0 @@
-38006F45291200
