--- 2021/day16-packet-decoder/test_input3.txt
@@ -1,1 +1,0 @@
-EE00D40C823060
