--- 2021/day16-packet-decoder/test_input4.txt
@@ -1,1 +1,0 @@
-8A004A801A8002F478
